Convert to Maven
Based primarily on work by markiewb [1, 2], migrate from Ant build to
Maven.

I was not able to get the OpenIDE-Module-Display-Category and
OpenIDE-Module-Short-Description properties in Bundle.properties to take
effect after moving the file to src/main/resources/<package>/. Looking
at several of markiewb's Maven based NetBeans plugins, he has had the
same problem and has defaulted to defining these properties directly in
the manifest instead -- that works, and we are not concerned with
localisation anyway.

To build the nbm:

    mvn package -Dkeystorepass=$password

[1] http://wiki.netbeans.org/DevFaqMavenHowToMigrateFromANT
[2] https://github.com/markiewb?tab=repositories